re-arrange lock management methods
